requirement.txt
bcrypt==4.3.0
beautifulsoup4==4.13.4
bs4==0.0.2
certifi==2025.4.26
charset-normalizer==3.4.2
dnspython==2.7.0
idna==3.10
numpy==2.2.6
pandas==2.2.3
pymongo==4.13.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.3
six==1.17.0
soupsieve==2.7
SQLAlchemy==2.0.41
tqdm==4.67.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
asyncpg
psycopg2-binary
pydantic
fastapi
fastapi[standard]
schemas
python-ulid
pyjwt
redis
